🪒 Razor: Flatten module hierarchies and optimize out .push_str(&format!(...))
- Flattened `src/errors/assembly.rs` into `src/errors/mod.rs` and deleted the former.
- Flattened `src/semantic/assembly/` folder by moving `model.rs` into `assembly.rs` and deleting the folder.
- Replaced multiple intermediate string allocations `foo.push_str(&format!(...))` with `write!(foo, ...)` in `src/tools/alchemist.rs`, `src/tools/weave.rs`, and `src/tools/labyrinth.rs`.
- Cleaned up workspace files.
